requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.7.3
cryptography==2.1.4
cycler==0.10.0
Django==1.11.20
django-filter==2.0.0
djangorestframework==3.9.2
enum34==1.1.6
futures==3.2.0
html5lib==1.0.1
idna==2.6
ipaddress==1.0.17
isort==4.3.15
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
mercurial==4.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pyflakes==1.6.0
pygobject==3.26.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyxdg==0.25
reportlab==3.5.13
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.3
webencodings==0.5.1
wrapt==1.11.1

